pyproject.toml
[project]
name = "walkai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.117.1",
    "pydantic[email]>=2.11.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.34",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a21",
]
